DecentralizedAI/src/utils/legal_compliance.py

# legal_compliance.py

from django.db import models

class LegalCompliance(models.Model):
    user = models.ForeignKey('auth.User', on_delete=models.CASCADE)
    is_compliant = models.BooleanField(default=False)
    compliance_date = models.DateTimeField(auto_now_add=True)

    def check_compliance(self):
        # Check user compliance with legal requirements
        # This is a placeholder function, actual implementation will depend on the specific legal requirements
        pass

    def update_compliance_status(self, status):
        self.is_compliant = status
        self.save()

    def get_compliance_status(self):
        return self.is_compliant
