DecentralizedAI/tests/utils/test_legal_compliance.py

import unittest
from src.utils import legal_compliance

class TestLegalCompliance(unittest.TestCase):

    def setUp(self):
        self.legal_compliance = legal_compliance.LegalCompliance()

    def test_check_compliance(self):
        result = self.legal_compliance.check_compliance()
        self.assertTrue(result)

    def test_update_compliance(self):
        result = self.legal_compliance.update_compliance()
        self.assertTrue(result)

    def test_get_compliance_status(self):
        result = self.legal_compliance.get_compliance_status()
        self.assertIsNotNone(result)

if __name__ == '__main__':
    unittest.main()

This is a basic test suite for the legal_compliance module. It tests the main functions: check_compliance, update_compliance, and get_compliance_status.